+++ custom_components/saj_modbus/text.py
@@ -1,4 +1,4 @@
-"""Platform for writable SAJ Modbus time entities."""
+"""Plattform für schreibbare SAJ Modbus Uhrzeit-Entitäten."""
 import datetime
 import re
 import logging
@@ -17,7 +17,7 @@ async def async_setup_entry(
     entry: ConfigEntry,
     async_add_entities: AddEntitiesCallback,
 ) -> None:
-    """Set up the writable time entities for First Charge."""
+    """Richte die schreibbaren Uhrzeit-Entitäten für First Charge ein."""
     hub = hass.data[DOMAIN][entry.entry_id]["hub"]
     entities = [
         SajFirstChargeStartTimeTextEntity(hub),
@@ -26,33 +26,33 @@ async def async_setup_entry(
     async_add_entities(entities)
 
 class SajFirstChargeStartTimeTextEntity(TextEntity):
-    """Writable time entity for the First Charge Start Time (Format HH:MM)."""
+    """Schreibbare Uhrzeit-Entität für den First Charge Start Time (Format HH:MM)."""
 
     def __init__(self, hub):
-        """Initialize the entity."""
+        """Initialisiere die Entität."""
         self._hub = hub
         self._attr_name = "SAJ First Charge Start Time (Time)"
         self._attr_unique_id = "saj_first_charge_start_time_time"
         self._attr_native_value = "00:00"
-        # Regex enforcing HH:MM format: hours from 00 to 23, minutes from 00 to 59
+        # Regex, das HH:MM erzwingt: Stunden von 00 bis 23, Minuten von 00 bis 59
         self._attr_pattern = r"^(0[0-9]|1[0-9]|2[0-3]):([0-5][0-9])$"
         self._attr_mode = "text"
 
     async def async_update(self) -> None:
-        """Update is not used here to avoid additional Modbus requests."""
-        # We intentionally leave this update block empty,
-        # so that Modbus registers are not queried again.
+        """Update wird hier nicht genutzt, um zusätzliche Modbus-Anfragen zu vermeiden."""
+        # Wir verlassen diesen Update-Block bewusst leer,
+        # damit hier nicht erneut modbus-Register abgefragt werden.
         pass
 
     async def async_set_value(self, value) -> None:
-        """Set a new start time value (Format 'HH:MM')."""
-        # If value is a datetime.time object, convert it
+        """Setze einen neuen Startzeitwert (Format 'HH:MM')."""
+        # Falls value ein datetime.time-Objekt ist, konvertiere es
         if isinstance(value, datetime.time):
             value = value.strftime("%H:%M")
         
         if not isinstance(value, str) or not re.match(self._attr_pattern, value):
             _LOGGER.error(
-                "Invalid time format for First Charge Start Time: %s. Expected HH:MM", value
+                "Ungültiges Zeitformat für First Charge Start Time: %s. Erwartet HH:MM", value
             )
             return
 
@@ -61,10 +61,10 @@ class SajFirstChargeStartTimeTextEntity(TextEntity):
         self.async_write_ha_state()
 
 class SajFirstChargeEndTimeTextEntity(TextEntity):
-    """Writable time entity for the First Charge End Time (Format HH:MM)."""
+    """Schreibbare Uhrzeit-Entität für den First Charge End Time (Format HH:MM)."""
 
     def __init__(self, hub):
-        """Initialize the entity."""
+        """Initialisiere die Entität."""
         self._hub = hub
         self._attr_name = "SAJ First Charge End Time (Time)"
         self._attr_unique_id = "saj_first_charge_end_time_time"
@@ -73,18 +73,18 @@ class SajFirstChargeEndTimeTextEntity(TextEntity):
         self._attr_mode = "text"
 
     async def async_update(self) -> None:
-        """Update is not used here to avoid additional Modbus requests."""
+        """Update wird hier nicht genutzt, um zusätzliche Modbus-Anfragen zu vermeiden."""
         pass
 
     async def async_set_value(self, value) -> None:
-        """Set a new end time value (Format 'HH:MM')."""
-        # If value is a datetime.time object, convert it
+        """Setze einen neuen Endzeitwert (Format 'HH:MM')."""
+        # Falls value ein datetime.time-Objekt ist, konvertiere es
         if isinstance(value, datetime.time):
             value = value.strftime("%H:%M")
         
         if not isinstance(value, str) or not re.match(self._attr_pattern, value):
             _LOGGER.error(
-                "Invalid time format for First Charge End Time: %s. Expected HH:MM", value
+                "Ungültiges Zeitformat für First Charge End Time: %s. Erwartet HH:MM", value
             )
             return
